# swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080/api",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/contact/default/{contactId}" : {
      "put" : {
        "tags" : [ "联系人管理" ],
        "summary" : "设置默认联系人",
        "description" : "将指定联系人设置为默认乘车人",
        "operationId" : "setDefault",
        "parameters" : [ {
          "name" : "contactId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/contact/cancel/default/{contactId}" : {
      "post" : {
        "tags" : [ "联系人管理" ],
        "summary" : "取消默认联系人",
        "description" : "取消指定联系人的默认状态",
        "operationId" : "cancelDefault",
        "parameters" : [ {
          "name" : "contactId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/contact/add" : {
      "post" : {
        "tags" : [ "联系人管理" ],
        "summary" : "添加联系人",
        "description" : "新增乘客联系人信息",
        "operationId" : "addContact",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ContactDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/ai/ticket/consult" : {
      "post" : {
        "tags" : [ "AI助手" ],
        "summary" : "票务咨询",
        "description" : "针对特定车次和日期的票务问题解答",
        "operationId" : "ticketConsult",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "additionalProperties" : {
                  "type" : "string"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ai/parse" : {
      "post" : {
        "tags" : [ "AI助手" ],
        "summary" : "解析用户输入",
        "description" : "将自然语言解析为结构化查询参数（出发地、目的地、日期、时间段、偏好）",
        "operationId" : "parseQuery",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "additionalProperties" : {
                  "type" : "string"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultMapStringString"
                }
              }
            }
          }
        }
      }
    },
    "/ai/order/consult" : {
      "post" : {
        "tags" : [ "AI助手" ],
        "summary" : "订单咨询",
        "description" : "针对特定订单的咨询问题解答",
        "operationId" : "orderConsult",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "additionalProperties" : {
                  "type" : "string"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ai/chat" : {
      "post" : {
        "tags" : [ "AI助手" ],
        "summary" : "统一聊天接口",
        "description" : "带会话管理的AI聊天，支持自动查票和多轮对话",
        "operationId" : "chat",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AiChatRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultAiChatResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/ai/chat/legacy" : {
      "post" : {
        "tags" : [ "AI助手" ],
        "summary" : "通用对话接口",
        "description" : "简单的AI对话，无会话管理功能",
        "operationId" : "chatLegacy",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "additionalProperties" : {
                  "type" : "string"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/contact/list/{userId}" : {
      "get" : {
        "tags" : [ "联系人管理" ],
        "summary" : "获取联系人列表",
        "description" : "获取指定用户的所有常用联系人",
        "operationId" : "getContacts",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultListContact"
                }
              }
            }
          }
        }
      }
    },
    "/ai/history/{sessionId}" : {
      "get" : {
        "tags" : [ "AI助手" ],
        "summary" : "获取聊天历史",
        "description" : "获取指定会话ID的聊天记录",
        "operationId" : "getHistory",
        "parameters" : [ {
          "name" : "sessionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultListChatMessageDTO"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "AI助手" ],
        "summary" : "清除聊天历史",
        "description" : "删除指定会话ID的聊天记录",
        "operationId" : "clearHistory",
        "parameters" : [ {
          "name" : "sessionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/contact/delete/{contactId}" : {
      "delete" : {
        "tags" : [ "联系人管理" ],
        "summary" : "删除联系人",
        "description" : "删除指定联系人ID的联系人",
        "operationId" : "deleteContact",
        "parameters" : [ {
          "name" : "contactId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "ResultBoolean" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "LocalTime" : {
        "type" : "object",
        "properties" : {
          "hour" : {
            "type" : "integer",
            "format" : "int32"
          },
          "minute" : {
            "type" : "integer",
            "format" : "int32"
          },
          "second" : {
            "type" : "integer",
            "format" : "int32"
          },
          "nano" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "TicketInfoDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "trainId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "trainNumber" : {
            "type" : "string"
          },
          "trainType" : {
            "type" : "string"
          },
          "departureStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "departureStationName" : {
            "type" : "string"
          },
          "arrivalStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "arrivalStationName" : {
            "type" : "string"
          },
          "departureDate" : {
            "type" : "string",
            "format" : "date"
          },
          "departureTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "arrivalTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "seatType" : {
            "type" : "string"
          },
          "seatTypeName" : {
            "type" : "string"
          },
          "availableCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "price" : {
            "type" : "number"
          },
          "priceText" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "ContactDTO" : {
        "type" : "object",
        "properties" : {
          "realName" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "idCard" : {
            "type" : "string"
          },
          "type" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ResultMapStringString" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "AiChatRequestDTO" : {
        "required" : [ "message" ],
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string"
          },
          "sessionId" : {
            "type" : "string"
          },
          "autoQuery" : {
            "type" : "boolean"
          }
        }
      },
      "AiChatResponseDTO" : {
        "type" : "object",
        "properties" : {
          "sessionId" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "params" : {
            "$ref" : "#/components/schemas/TicketQueryParamDTO"
          },
          "tickets" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TicketInfoDTO"
            }
          },
          "paramMissingCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "missingParams" : {
            "type" : "string"
          }
        }
      },
      "ResultAiChatResponseDTO" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AiChatResponseDTO"
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "TicketQueryParamDTO" : {
        "type" : "object",
        "properties" : {
          "from" : {
            "type" : "string"
          },
          "to" : {
            "type" : "string"
          },
          "date" : {
            "type" : "string"
          },
          "timeRange" : {
            "type" : "string"
          },
          "preference" : {
            "type" : "string"
          }
        }
      },
...
      "Contact" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "idCard" : {
            "type" : "string"
          },
          "type" : {
            "type" : "integer",
            "format" : "int32"
          },
          "isDefault" : {
            "type" : "integer",
            "format" : "int32"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updateTime" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ResultListContact" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Contact"
            }
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "ChatMessageDTO" : {
        "type" : "object",
        "properties" : {
          "role" : {
            "type" : "string"
          },
          "content" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ResultListChatMessageDTO" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ChatMessageDTO"
            }
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "ResultVoid" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "object"
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      }
    }
  }
}